Print: Use point sizes
Using inches was not even right in a philosophical sense, because
the size calculated by the renderer already is in points.
Now the output is sized like PowerPoint does it, and using integral
point values gets rid of the annoying white lines on the bottom
of the printed page (looks like PDF can handle integral sizes only).