Parser: add specialized EmK mode
The EmK mode can be specified with a '.mode emk' command, or,
if no mode is specified, it is autodetected based on the
presence of EmK-Style constructs found in the document.

When in EmK mode, the parse now makes an effort to detect
the order based on refrain markers present in the verses.